Generic function cmp
Data Frames will be arrays of series, so every series have to  be
of sae type, while storing diferent basic types. To prevent having
to export one function for each basic type of series, we'll use
generic functions like dfl_cmp which will be able to perform
comparisons between arguments of different types. This functions
will accesi an array of function pointer indexed by enum of basic types,
so it will switch to the correct function.